requirements.txt
# RAG Comparison - Requirements
# Install with: pip install -r requirements.txt

# Core LLM
openai>=1.0.0

# RAG Frameworks
langchain>=0.3.0               # note: versioning jumped to 1.x, >=0.3 still resolves correctly
langchain-openai>=0.3.0       # updated: was >=0.2.0, current is 1.1.10
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-mistralai>=0.2.12
llama-index>=0.11.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-openai>=0.3.0
dspy>=2.5.0                    # v3.x API — do not downgrade to dspy-ai

# Vector store
chromadb>=0.5.0,<2.0.0         # updated: current is 1.5.1, cap at <2.0.0 to avoid future breaking changes
faiss-cpu>=1.8.0

# Evaluation
ragas==0.4.3                   # pinned: 0.4.x API is what our synthetic_data.py is written against
datasets>=3.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
jupyter>=1.0.0
ipykernel>=6.29.0